requirements.txt
discord.py
gTTS
Pillow
aiohttp
nest_asyncio
unidecode